Sort list elements in place
appendChild() method actually moves the element from one place to
another in the DOM, not copying it.
So when we call appendChild() on each item in the sorted items array,
we simply move each item from its current location to a new location
in the list, so there are no duplicate items.